// src/app/model/ubigeo.ts
export class Ubigeo {
    idUbigeo:   number;
    ubipais:          string;
    ubidpto:          string;
    ubiprovincia:       string;
    ubidiscrito:        string;
    denominacion:   string;
    longitud:          string;
    latitud:  string;
    poblacion: number;
    superficie:        string;
    ubiestado: number;
}